DDCNL-10523 scala 3 and upgrade dependencies

+++ app/repository/MinimalDetailsCacheRepository.scala
@@ -35,8 +35,9 @@ package repository
 import com.google.inject.Inject
 import com.mongodb.client.model.FindOneAndUpdateOptions
 import crypto.DataEncryptor
-import org.mongodb.scala.model._
-import play.api.libs.json._
+import org.mongodb.scala.gridfs.{ObservableFuture, SingleObservableFuture}
+import org.mongodb.scala.model.*
+import play.api.libs.json.*
 import play.api.{Configuration, Logging}
 import uk.gov.hmrc.mongo.MongoComponent
 import uk.gov.hmrc.mongo.play.json.formats.MongoJavatimeFormats
@@ -62,7 +63,7 @@ class MinimalDetailsCacheRepository @Inject()(
         keys = Indexes.ascending("lastUpdated"),
         indexOptions = IndexOptions()
           .name("dataExpiry")
-          .expireAfter(config.get[Int](path = "mongodb.minimal-detail.timeToLiveInSeconds"), TimeUnit.SECONDS)
+          .expireAfter(config.get[Int](path = "mongodb.minimal-detail.timeToLiveInSeconds").toLong, TimeUnit.SECONDS)
           .background(true)
       ),
       IndexModel(
